refactor: move Main & 404 to pages folder, change husky hook
