+++ RS_React/.husky/pre-push
@@ -1,4 +1,4 @@
 #!/usr/bin/env sh
 . "$(dirname -- "$0")/_/husky.sh"
 exec >/dev/tty 2>&1
-cd ./RS_React && npm run husky-test
+cd ./RS_React && npm run pre-push
