+++ RS_React/src/pages/Main/Main.tsx
@@ -1,5 +1,5 @@
 import { Outlet } from 'react-router-dom';
-import Wrapper from '../Wrapper/Wrapper';
+import Wrapper from '../../components/Wrapper/Wrapper';
 import './Main.css';
 
 export default function Main() {
